Move pkg/api to core/api
