+++ mgmt/rest/server.go
@@ -32,9 +32,9 @@ import (
 	"github.com/julienschmidt/httprouter"
 	"github.com/urfave/negroni"
 
+	"github.com/intelsdi-x/snap/core/api"
 	"github.com/intelsdi-x/snap/mgmt/rest/v1"
 	"github.com/intelsdi-x/snap/mgmt/rest/v2"
-	"github.com/intelsdi-x/snap/pkg/api"
 )
 
 var (
